Move P2PDumpManager
