+++ infrastructure/logging/build.gradle
@@ -6,8 +6,6 @@ dependencies {
     implementation 'org.apache.logging.log4j:log4j-core'
     compileOnly 'com.github.oshi:oshi-core'
 
-    testImplementation testFixtures(project(':ethereum:spec'))
-    testImplementation testFixtures(project(':infrastructure:time'))
     testFixturesImplementation 'org.apache.logging.log4j:log4j-core'
     testImplementation 'com.github.oshi:oshi-core'
 }
