+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/P2PDumpManager.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.infrastructure.logging;
+package tech.pegasys.teku.networking.eth2;
 
 import java.io.IOException;
 import java.nio.file.FileAlreadyExistsException;
